PART 2.4 & 2.5 separate no count csv and then scrape version 2 (direct airdrop)
reddit_sentiment_analysis_totalcount01.py
2.4.
Because realize is it not sufficient, use the same method, separate those that have zero count again in the total column and then save as another csv file and carry on to scrape version 2 method that was did.

2.5.
ver2_reddit_crawl_airdrop_name.py (use back existing code)
Carry on scrape directly the airdrop name at reddit  and see if there is any comments out. Just need to change the file directory of the csv to run.