All the previous model comparison version1
All these techniques did not actually output well in the f1, precision and recall scores, need to fine tune the threshold quite badly which is not recommended in the first place too.

Although there are certain useful results, this will be use to present the challenges faced if needed to come out something for sentiment analysis portion.