Included fixing imbalance class of data and perform further cross validation model comparison script & results of bert as well as fixing the threshold & the combine of the csv models too if it is split to different parts
initial threshold was not same from the initial script when analyzing, so further combine and tune the threshold to same as the rest of the model too, 0.01.

Heads up, the combine_roberta_sentiment_analysis_results.py can be use for any of the model to combine the csv file because i split the analysis to 2-3 different parts as it take super long and scared the script auto close or stop.